backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.66
python-dotenv==1.0.1
pydantic-settings==2.6.1
